moved interval logic around
