+++ armor-write/src/test/java/com/rapid7/armor/FileStoreV2Test.java
@@ -1,6 +1,6 @@
 package com.rapid7.armor;
 
-import static com.rapid7.armor.Constants.MAX_INTERVAL;
+import static com.rapid7.armor.schema.Interval.MAX_INTERVAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
