+++ armor-write/src/test/java/com/rapid7/armor/S3StoreTest.java
@@ -32,7 +32,7 @@ import java.util.UUID;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import static com.rapid7.armor.Constants.MAX_INTERVAL;
+import static com.rapid7.armor.schema.Interval.MAX_INTERVAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
